+++ routes/b2b.php
@@ -93,14 +93,14 @@ Route::middleware('validate.header')
 
                     // Product
                     Route::prefix('product')->group(function (): void {
-                        Route::get('/', 'getAllProduct');
                         Route::post('/add', 'addProduct');
                         Route::get('/analytic/{user_id}', 'getAnalytics');
-                        Route::get('/details/{product_id}', 'getProductById');
+                        Route::get('/details/{product_id}/{user_id}', 'getProductById');
                         Route::post('/update', 'updateProduct');
-                        Route::delete('/delete/{product_id}', 'deleteProduct');
+                        Route::delete('/delete/{user_id}/{product_id}', 'deleteProduct');
                         Route::post('import', 'productImport');
                         Route::post('export/{user_id}/{type}', 'export');
+                        Route::get('/{user_id}', 'getAllProduct');
                     });
 
                     // Shipping
